covered_calls_analysis/requirements.txt
pandas>=2.1.0
numpy>=1.25.0
matplotlib>=3.7.0
seaborn>=0.13.2
pyyaml>=6.0
requests>=2.31.0
pytz>=2023.3
